Modify CMakeLists.txt file to use customized Oculus OpenXR loader
OpenXR is configured to build using a standard Android loader, which apparently did not appear to work with the Oculus Quest.  We need to use the one from Oculus to get this to work.

Based on instructions in: https://gayanediriweera.github.io/code/2021/04/06/how-to-run-helloxr-on-oculus-quest.html

(I've tried to use a relative path here to avoid needing to set up an environment variable.  If it didn't, you'll see a later commit to fix it.)